fix : extract IProps from typing file
