+++ src/components/Button/Button.tsx
@@ -1,9 +1,34 @@
-import { FC } from 'react'
+import { ButtonHTMLAttributes, FC, MouseEventHandler } from 'react'
+import { Icon } from 'react-feather'
 
 import { TestingProps } from '@/testing'
 
 import { createIcon, getClassNames } from './internal/Button.helpers'
-import { IProps } from './internal/Button.types'
+import { IButtonColor, IButtonStyle } from './internal/Button.types'
+
+interface IPropsBase {
+  className?: string
+  text: string
+  onClick?: MouseEventHandler | never
+  color?: IButtonColor
+  style?: IButtonStyle
+  small?: boolean
+  iconStart?: Icon
+  iconEnd?: Icon
+  type?: ButtonHTMLAttributes<HTMLButtonElement>['type']
+}
+
+interface IPropsSubmit extends IPropsBase {
+  type: 'submit'
+  onClick?: never
+}
+
+interface IPropsButton extends IPropsBase {
+  type?: 'button' | 'reset'
+  onClick: MouseEventHandler
+}
+
+type IProps = IPropsButton | IPropsSubmit
 
 export const Button: FC<IProps & TestingProps> = ({
   text,
@@ -17,18 +42,16 @@ export const Button: FC<IProps & TestingProps> = ({
   'data-testid': dataTestid,
   type = 'button',
 }) => {
-  const classNames = getClassNames({ color, style, small, className })
-
   return (
     <button
       onClick={onClick}
-      className={classNames.join(' ')}
+      className={getClassNames(color, style, small, className)}
       data-testid={dataTestid}
       type={type}
     >
-      {iconStart && createIcon({ icon: iconStart, small, position: 'start' })}
+      {iconStart && createIcon(iconStart, small, 'start')}
       {text}
-      {iconEnd && createIcon({ icon: iconEnd, small, position: 'end' })}
+      {iconEnd && createIcon(iconEnd, small, 'end')}
     </button>
   )
 }
