+++ src/components/Button/internal/Button.types.ts
@@ -1,26 +1,5 @@
-import { ButtonHTMLAttributes, MouseEventHandler } from 'react'
-import { Icon } from 'react-feather'
+import { IColors } from '@/typing/colors'
 
-interface IPropsBase {
-  className?: string
-  text: string
-  onClick?: MouseEventHandler | never
-  color?: 'primary' | 'white'
-  style?: 'full' | 'outlined'
-  small?: boolean
-  iconStart?: Icon
-  iconEnd?: Icon
-  type?: ButtonHTMLAttributes<HTMLButtonElement>['type']
-}
+export type IButtonColor = Extract<IColors, 'base-white' | 'primary'>
 
-export interface IPropsSubmit extends IPropsBase {
-  type: 'submit'
-  onClick?: never
-}
-
-export interface IPropsButton extends IPropsBase {
-  type?: 'button' | 'reset'
-  onClick: MouseEventHandler
-}
-
-export type IProps = IPropsButton | IPropsSubmit
+export type IButtonStyle = 'full' | 'outlined'
